Allow physical birth time bigger than logical
During regular block writes ZFS sets both logical and physical
birth times equal to the current TXG.  During dedup and block
cloning logical birth time is still set to the current TXG, but
physical may be copied from the original block that was used.
This represents the fact that logically user data has changed,
but the physically it is the same old block.

But block rewrite introduces a new situation, when block is not
changed logically, but stored in a different place of the pool.
From ARC, scrub and some other perspectives this is a new block,
but for example for user applications or incremental replication
it is not.

This patch prepares the ground for the functionality of physical-
only rewrite by allowing ZFS to handle block pointers with physical
birth time bigger than logical.  The new functionality is not used
at this point, and the only expected change is that error log is
now accounted in terms of physical birth times, rather than logical,
since if a block with logged error was somehow rewritten then the,
previous error does not matter any more.

This change also introduces a new TRAVERSE_LOGICAL flag to traverse
code, allowing zfs send, redact and diff to work in context of
logical birth times, ignoring physical-only rewrites.  It also
changes nothing at this point due to lack of those changes, but
they will come in a following patch.

Signed-off-by:	Alexander Motin <mav@FreeBSD.org>
Sponsored by:	iXsystems, Inc.